Extracted a few class/interface files that were embedded in other files
Extracted a few class/interface files that were embedded in other files
